workspace/Python4_Homework01/src/test_composable.py
'''
Created on Nov 7, 2015

@author: prefalo
'''
import unittest
from composable import Composable

def reverse(s):
    "Reverses a string using negative-stride sequencing."
    return s[::-1]

def square(x):
    "Multiplies a number by itself."
    return x*x

class ComposableTestCase(unittest.TestCase):

    def test_inverse(self):
        reverser = Composable(reverse)
        nulltran = reverser * reverser
        for s in "", "a", "0123456789", "abcdefghijklmnopqrstuvwxyz":
            self.assertEqual(nulltran(s), s)
    
    def test_square(self):
        squarer = Composable(square)
        po4 = squarer * squarer
        for v, r in ((1, 1), (2, 16), (3, 81)):
            self.assertEqual(po4(v), r)
            
    def test_exceptions(self):
        fc = Composable(square)
        with self.assertRaises(TypeError):
            fc = fc * 3
        with self.assertRaises(TypeError):
            fc = square * fc
            
    def test_pow(self):
        squarer = Composable(square)
        po = squarer ** 2
        for v, r in ((0, 0), (1, 1), (2, 16), (3, 81), (4, 256)):
            self.assertEqual(po(v), r)
        for a, b in ((1, 5), (6, 6)):
            self.assertNotEqual(po(a), b)
   

   


if __name__ == "__main__":
    #import sys;sys.argv = ['', 'Test.testName']
    unittest.main()